MEDIUM: h1: Immediately try to read data for frontend
In h1_init(), if we're a frontend connection, immediately attempt to
read data, if the connection is ready, instead of just subscribing.
There may already be data available, at least if we're using 0RTT.

This may be backported up to 2.8 in a while, after 3.3 is released, so
that if it causes problem, we have a chance to hear about it.

(cherry picked from commit b3d6f44af8509063bf98591b4b86e9d64e73a8c1)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>