BUG/MEDIUM: mux-h2: make sure not to move a dead connection to idle
In h2_detach(), it looks possible to place a dead connection back to
the idle list, and to later call h2_release() on it once detected as
dead. It's not certain that it happens but nothing in the code shows
it is not possible, so better make sure it cannot happen.

This should be preventively backported to all versions.

(cherry picked from commit 59c599f3f07c549c195f1bd877d2c2a09e6bc61f)
Signed-off-by: Willy Tarreau <w@1wt.eu>