BUG/MINOR: quic: close connection on CID alloc failure
During RETIRE_CONNECTION_ID frame parsing, a new connection ID is
immediately reallocated after the release of the previous one. This is
done to ensure that the peer will never run out of DCID.

Prior to this patch, a CID allocation failure was be silently ignored.
This prevent the emission of a new CID, which could prevent the peer to
emit packets if it had no other CIDs available for use. Now, such error
is considered fatal to the connection. This is the safest solution as
it's better to close connections when memory is running low.

It must be backported up to 2.8.

(cherry picked from commit 73621adb23688d421c43df42f85f5171c2b3eafe)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit cf0ef486cff51fa32dbabca198808a395d306d17)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>