MINOR: quic: adjust CID conn tree alloc in qc_new_conn()
Change qc_new_conn() so that the connection CID tree is allocated
earlier in the function. This patch does not introduce a behavior
change. Its objective is to facilitate future evolutions on CIDs
handling.

This patch is a prerequisite for the fix on CID collision, thus it must
be backported prior to it to every affected version.

(cherry picked from commit 0ef473ba6b3f5b555aec1b3ef93debb0db3660ac)
[ad: context adjusted]
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>