BUG/MEDIUM: stick-tables: Make sure updates are seen as local
In stktable_touch_with_exp, if it is a local update, add it to the
pending update list even if it's already in the tree as a remote update,
otherwise it will never be communicated to other peers;
It used to work before 3.2 because of the ordering of operations, but
it's been broken by adding an extra step with the pending update list,
so we now have to explicitely check for that.

This should be backported to 3.2.

(cherry picked from commit c26bcfc1e3ae0193e2589a15c96ffd44b2e0f770)
Signed-off-by: Willy Tarreau <w@1wt.eu>