Revert "BUG/MEDIUM: connections: permit to permanently remove an idle conn"
The target patch fixes a rare race condition which happen when a MUX IO
handler is working on a connection already moved into the purge list. In
this case, the handler will incorrectly moved back the connection into
the idle list.

To fix this, conn_delete_from_tree() was extended to remove flags along
with the connection from the idle list. This was performed when the
connection is moved into the purge list. However, it introduces another
issue related to the idle server connection accounting. Thus it is
necessary to revert it prior to the incoming newer fix.

This patch must be backported to every version where the original commit
is.

(cherry picked from commit d79295d89bc086caa5ddf7ef255c711429a97174)
[ad: conflict due to extra tid argument conn_delete_from_tree() not
 available in 3.2]
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>