BUG/MINOR: acme: alert when the map doesn't exist at startup
When configuring an acme section with the 'map' keyword, the user must
use an existing map. If the map doesn't exist, a log will be emitted
when trying to add the challenge to the map.

This patch change the behavior by checking at startup if the map exists,
so haproxy would warn and won't start with a non-existing map.

This must be backported in 3.2.

(cherry picked from commit e0665d4ffe1b382d2fccec3c5bb57a9ff1de2351)
Signed-off-by: Willy Tarreau <w@1wt.eu>