MINOR: config: Do proto detection for listeners before checks about ALPN
The verification of any forced mux protocol, via the "proto" keyword, for
listeners is now performed before any tests on the ALPN. It will be
mandatory to be able to force the default ALPN, if not forced on the bind
line.

This patch will be mandatory for the next fix.

(cherry picked from commit 2ef8b91a0013df7f9c5c3a8e4b4028863a61afae)
Signed-off-by: Willy Tarreau <w@1wt.eu>