BUG/MEDIUM: applet: Fix conditions to detect spinning loop with the new API
There was a mixup between read/send events and ability for an applet to
receive and send. The fix seems obvious by reading it. The call-rate must be
incremented when nothing was received from the applet while it was allowed
and nothing was sent to the applet while it was allowed.

This patch must be backported as far as 3.0.

(cherry picked from commit dfdccbd2af56a9cb165037d106f149cdd4eb8c80)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>