MINOR: quic: implement cc-algo server keyword
Extend QUIC server configuration so that congestion algorithm and
maximum window size can be set on the server line. This can be achieved
using quic-cc-algo keyword with a syntax similar to a bind line.

This should be backported up to 3.3 as this feature is considered as
necessary for full QUIC backend support. Note that this relies on the
serie of previous commits which should be picked first.

(cherry picked from commit 47dff5be52dcfdbb02024f658c9cff0f1ca31e8c)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>